+++ CHANGES.rst
@@ -7,6 +7,7 @@ Here you can see the full list of changes between each SQLAlchemy-Continuum rele
 1.0-b5 (2014-04-xx)
 ^^^^^^^^^^^^^^^^^^^
 
+- Added order_by mapper arg ignoring for version class reflection if other than string argument is used
 - SQLAlchemy-Utils dependency updated to 0.25.4